feat(client): superclass canvas rendering
